added authorization

+++ backend/routes/posts.js
@@ -27,12 +27,14 @@ const storage = multer.diskStorage({
   }
 });
 
-router.post("", multer({storage: storage}).single("image"), (req, res, next) => {
+router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
   const url = req.protocol + "://" + req.get("host");
+  const userData = req.userData;
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" +req.file.filename
+    imagePath: url + "/images/" +req.file.filename,
+    creator: userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -68,14 +70,20 @@ router.get("", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then((response) => {
-    res.status(200).json({
-      message: 'Post deleted'
-    });
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then((response) => {
+    if(response.n === 0) {
+      res.status(401).json({
+        message: 'Not authorise'
+      });
+    } else {
+      res.status(200).json({
+        message: 'Post deleted'
+      });
+    }
   });
 });
 
-router.put("/:id",checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
+router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
   let imagePath = "";
   if (req.file) {
     const url = req.protocol + "://" + req.get("host");
@@ -87,14 +95,20 @@ router.put("/:id",checkAuth, multer({storage: storage}).single("image"), (req, r
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.authData.userId
   });
-  console.log(post);
 
-  Post.updateOne({_id: req.params.id}, post).then((result) => {
-    res.status(200).json({
-      message: 'Post Updated'
-    });
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then((result) => {
+    if(result.nModified === 0) {
+      res.status(401).json({
+        message: 'Not authorise'
+      });
+    } else {
+      res.status(200).json({
+        message: 'Post Updated'
+      });
+    }
   });
 });
